AbstractParallelCollector
